feat: registration fields validators
