+++ playfit-mobile/playfit/lib/authentification/registration_step1.dart
@@ -1,0 +1,155 @@
+import 'package:flutter/material.dart';
+
+class RegistrationStep1 extends StatefulWidget {
+  final TextEditingController usernameController;
+  final TextEditingController emailController;
+  final TextEditingController passwordController;
+  final TextEditingController confirmPasswordController;
+
+  const RegistrationStep1({
+    super.key,
+    required this.usernameController,
+    required this.emailController,
+    required this.passwordController,
+    required this.confirmPasswordController,
+  });
+
+  @override
+  State<RegistrationStep1> createState() => _RegistrationStep1State();
+}
+
+class _RegistrationStep1State extends State<RegistrationStep1> {
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 40.0),
+      child: Column(
+        children: [
+          TextFormField(
+            controller: widget.usernameController,
+            decoration: InputDecoration(
+              hintText: 'Nom d\'utilisateur',
+              filled: true,
+              fillColor: Colors.white,
+              prefixIcon: const Icon(Icons.person),
+              suffixIcon: widget.usernameController.text.isNotEmpty
+                  ? IconButton(
+                      icon: const Icon(Icons.close),
+                      onPressed: () => widget.usernameController.clear(),
+                    )
+                  : null,
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(100.0),
+                borderSide: BorderSide.none,
+              ),
+            ),
+            validator: (value) {
+              if (value == null || value.isEmpty) {
+                return 'Veuillez entrer un nom d\'utilisateur';
+              }
+              if (value.length < 4) {
+                return 'Le nom d\'utilisateur doit contenir au moins 4 caractères';
+              }
+              return null;
+            },
+          ),
+          const SizedBox(height: 10),
+          TextFormField(
+            controller: widget.emailController,
+            decoration: InputDecoration(
+              hintText: 'Adresse e-mail',
+              filled: true,
+              fillColor: Colors.white,
+              prefixIcon: const Icon(Icons.email),
+              suffixIcon: widget.emailController.text.isNotEmpty
+                  ? IconButton(
+                      icon: const Icon(Icons.close),
+                      onPressed: () => widget.emailController.clear(),
+                    )
+                  : null,
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(100.0),
+                borderSide: BorderSide.none,
+              ),
+            ),
+            keyboardType: TextInputType.emailAddress,
+            validator: (value) {
+              if (value == null || value.isEmpty) {
+                return 'Veuillez entrer une adresse e-mail';
+              }
+              if (!RegExp(r'^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$')
+                  .hasMatch(value)) {
+                return 'Veuillez entrer une adresse e-mail valide';
+              }
+              return null;
+            },
+          ),
+          const SizedBox(height: 10),
+          TextFormField(
+            controller: widget.passwordController,
+            obscureText: true,
+            decoration: InputDecoration(
+              hintText: 'Mot de passe',
+              filled: true,
+              fillColor: Colors.white,
+              prefixIcon: const Icon(Icons.lock_outline),
+              suffixIcon: widget.passwordController.text.isNotEmpty
+                  ? IconButton(
+                      icon: const Icon(Icons.close),
+                      onPressed: () => widget.passwordController.clear(),
+                    )
+                  : null,
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(100.0),
+                borderSide: BorderSide.none,
+              ),
+            ),
+            validator: (value) {
+              if (value == null || value.isEmpty) {
+                return 'Veuillez entrer un mot de passe';
+              }
+              if (value.length < 8) {
+                return 'Le mot de passe doit contenir au moins 8 caractères';
+              }
+              if (!RegExp(r'^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]+$')
+                  .hasMatch(value)) {
+                return 'Le mot de passe doit contenir au moins une lettre et un chiffre';
+              }
+              return null;
+            },
+          ),
+          const SizedBox(height: 10),
+          TextFormField(
+            controller: widget.confirmPasswordController,
+            obscureText: true,
+            decoration: InputDecoration(
+              hintText: 'Confirmez le mot de passe',
+              filled: true,
+              fillColor: Colors.white,
+              prefixIcon: const Icon(Icons.lock_outline),
+              suffixIcon: widget.confirmPasswordController.text.isNotEmpty
+                  ? IconButton(
+                      icon: const Icon(Icons.close),
+                      onPressed: () => widget.confirmPasswordController.clear(),
+                    )
+                  : null,
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(100.0),
+                borderSide: BorderSide.none,
+              ),
+            ),
+            validator: (value) {
+              if (value == null || value.isEmpty) {
+                return 'Veuillez confirmer votre mot de passe';
+              }
+              if (value != widget.passwordController.text) {
+                return 'Les mots de passe ne correspondent pas';
+              }
+              return null;
+            },
+          ),
+        ],
+      ),
+    );
+  }
+}
